reorganizer class ini package
