+++ app/src/main/java/com/abahstudio/githubuserdicoding/adapter/SectionsPagerAdapter.kt
@@ -1,8 +1,10 @@
-package com.abahstudio.githubuserdicoding
+package com.abahstudio.githubuserdicoding.adapter
 
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.abahstudio.githubuserdicoding.fragment.FollowerFragment
+import com.abahstudio.githubuserdicoding.fragment.FollowingFragment
 
 class SectionsPagerAdapter(activity: AppCompatActivity) : FragmentStateAdapter(activity) {
     override fun getItemCount(): Int {
