+++ app/src/main/java/com/abahstudio/githubuserdicoding/fragment/FollowingFragment.kt
@@ -1,10 +1,11 @@
-package com.abahstudio.githubuserdicoding
+package com.abahstudio.githubuserdicoding.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.abahstudio.githubuserdicoding.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
